requirements.txt
pymupdf>=1.23.0
pdfplumber>=0.10.3
camelot-py[cv]>=0.11.0
pandas>=2.1.0
numpy>=1.26.0
pydantic>=2.7.0
pyyaml>=6.0.1
rich>=13.7.0
openai>=1.40.0
layoutparser>=0.3.4
torch>=2.2.0
transformers>=4.43.0
sentencepiece>=0.2.0
huggingface_hub>=0.23.0
pytesseract>=0.3.10
opencv-python>=4.9.0.80
python-bidi>=0.4.2
arabic-reshaper>=3.0.0
docling>=2.0.0
